New functionality
- add all functions into Movie class
- rename files from anime -> movie
- add config_verification() to get config information outside from config.json file, check it, and return result of verification
- add config.json as example (path can be with \\ or with /)
