+++ config.json
@@ -1,0 +1,5 @@
+{
+	"FFMPEG": "F:\\Program Files\\Jellyfin\\Server\\FFMPEG.exe",
+	"MOVIES_PATH": "E:\\Program Data\\Videos\\Movies\\Black.Lagoon",
+	"BASE_TEMPLATE": "Black.Lagoon"
+}
